+++ rx/concurrency/mainloopscheduler/tkinterscheduler.py
@@ -1,11 +1,11 @@
 import logging
-from datetime import datetime, timedelta
 
 from rx.disposables import Disposable, SingleAssignmentDisposable, \
     CompositeDisposable
 from rx.concurrency.scheduler import Scheduler
 
 log = logging.getLogger("Rx")
+
 
 class TkinterScheduler(Scheduler):
     """A scheduler that schedules work via the Tkinter main event loop.
@@ -49,6 +49,7 @@ class TkinterScheduler(Scheduler):
             return scheduler.schedule(action, state)
 
         disposable = SingleAssignmentDisposable()
+
         def interval():
             disposable.disposable = action(scheduler, state)
 
@@ -73,3 +74,31 @@ class TkinterScheduler(Scheduler):
 
         duetime = self.to_datetime(duetime)
         return self.schedule_relative(duetime - self.now(), action, state)
+
+    def schedule_periodic(self, period, action, state=None):
+        """Schedules a periodic piece of work to be executed in the tkinter
+        mainloop.
+
+        Keyword arguments:
+        period -- Period in milliseconds for running the work periodically.
+        action -- Action to be executed.
+        state -- [Optional] Initial state passed to the action upon the first
+            iteration.
+
+        Returns the disposable object used to cancel the scheduled recurring
+        action (best effort)."""
+
+        state = [state]
+
+        def interval():
+            state[0] = action(state[0])
+            alarm[0] = self.master.after(period, interval)
+
+        log.debug("timeout: %s", period)
+        alarm = [self.master.after(period, interval)]
+
+        def dispose():
+            # nonlocal alarm
+            self.master.after_cancel(alarm[0])
+
+        return Disposable(dispose)