+++ rx/testing/reactivetest.py
@@ -5,6 +5,7 @@ from rx.notification import OnNext, OnError, OnCompleted
 from .recorded import Recorded
 from .subscription import Subscription
 
+
 def is_prime(i):
     """Tests if number is prime or not"""
 
@@ -18,6 +19,7 @@ def is_prime(i):
 
     return True
 
+
 # New predicate tests
 class OnNextPredicate(object):
     def __init__(self, predicate):
@@ -26,11 +28,12 @@ class OnNextPredicate(object):
     def __eq__(self, other):
         if other == self:
             return True
-        if other == None:
+        if other is None:
             return False
         if other.kind != 'N':
             return False
         return self.predicate(other.value)
+
 
 class OnErrorPredicate(object):
     def __init__(self, predicate):
@@ -39,11 +42,12 @@ class OnErrorPredicate(object):
     def __eq__(self, other):
         if other == self:
             return True
-        if other == None:
+        if other is None:
             return False
         if other.kind != 'E':
             return False
         return self.predicate(other.exception)
+
 
 class ReactiveTest(object):
     created = 100
@@ -52,14 +56,14 @@ class ReactiveTest(object):
 
     @classmethod
     def on_next(cls, ticks, value):
-        if type(value) == types.FunctionType:
+        if isinstance(value, types.FunctionType):
             return Recorded(ticks, OnNextPredicate(value))
 
         return Recorded(ticks, OnNext(value))
 
     @classmethod
     def on_error(cls, ticks, exception):
-        if type(exception) == types.FunctionType:
+        if isinstance(exception, types.FunctionType):
             return Recorded(ticks, OnErrorPredicate(exception))
 
         return Recorded(ticks, OnError(exception))
@@ -71,4 +75,3 @@ class ReactiveTest(object):
     @classmethod
     def subscribe(cls, start, end):
         return Subscription(start, end)
-