+++ rx/testing/recorded.py
@@ -1,5 +1,6 @@
 from rx.internal.basic import default_comparer
 
+
 class Recorded(object):
     def __init__(self, time, value, comparer=None):
         self.time = time
@@ -18,5 +19,4 @@ class Recorded(object):
         return str(self)
 
     def __str__(self):
-        return  "%s@%s" % (self.value, self.time)
-
+        return "%s@%s" % (self.value, self.time)