+++ tests/test_observable/py3_fromfuture.py
@@ -1,6 +1,11 @@
 import unittest
-import asyncio
-from asyncio import Future
+
+from nose import SkipTest
+import rx
+asyncio = rx.config['asyncio']
+if asyncio is None:
+    raise SkipTest("asyncio not available")
+Future = rx.config['Future']
 
 from rx import Observable